dsc-drivers: update to 24.05.1-001 (#48)
Summary
  - rework of platform_device management to better follow kernel device model
  - refinements on the missed-doorbell workaround
  - whitespace and comment cleanup in ionic_if.h

Internal patch list:
    ionic: simplify alloc_netdev routines
    ionic: platform code cleanup
    drivers: complain about early probe no resources
    drivers: no platform devicetree match
    mdev: simplify the uio driver naming
    mdev: simplify resource builder function calls
    mdev: clean up devicetree node search
    mdev: rework for dynamic pdev and no direct probe
    ionic: no export of probe and remove
    drivers: fix up platform device name
    Makefile: add kernel ver message
    drivers: fix up code checker complaints
    mdev: depend on the is_uio bit for using UIO driver
    mdev: remove devices on rmmod
    ionic: correct use of platform_data
    ionic: remove missed doorbell per-queue timer
    ionic: add private workqueue per-device
    ionic: add work item for missed-doorbell check
    ionic: add per-queue napi_schedule for doorbell check
    ionic: check for queue deadline in doorbell_napi_work
    ionic: use cancel_delayed_work
    ionic: Use an u16 for rx_copybreak
    ionic: Only run the doorbell workaround for certain asic_type(s)
    ionic: Keep interrupt affinity up to date
    ionic: Replace last_cpu with ionic_get_preferred_cpu
    ionic: Use netdev_name() function instead of netdev->name
    ionic: Align with DIMLIB's new ability to be a loadable module
    ionic: invoke softirqs after __napi_schedule in non-interrupt context
    ionic_if: fix up code style issues

+++ drivers/linux/eth/ionic/ionic_main.c
@@ -23,8 +23,8 @@ MODULE_LICENSE("GPL");
 MODULE_VERSION(IONIC_DRV_VERSION);
 MODULE_INFO(supported, "external");
 
-unsigned int rx_copybreak = IONIC_RX_COPYBREAK_DEFAULT;
-module_param(rx_copybreak, uint, 0600);
+unsigned short rx_copybreak = IONIC_RX_COPYBREAK_DEFAULT;
+module_param(rx_copybreak, ushort, 0600);
 MODULE_PARM_DESC(rx_copybreak, "Maximum size of packet that is copied to a bounce buffer on RX");
 
 unsigned int rx_fill_threshold = IONIC_RX_FILL_THRESHOLD;
